+++ core/src/main/java/org/simpledb/table/reader/Reader.java
@@ -1,8 +1,5 @@
 package org.simpledb.table.reader;
 
-import static org.simpledb.utils.UnsafeUtils.unsafe;
-
-
 /**
  * Created by yvladimirov on 9/25/15.
  */
@@ -10,6 +7,6 @@ public interface Reader<T extends Comparable> {
 
     public T read(long address);
 
-    long getSize(long ptr);
+    int getSize(long ptr);
 
 }
